Login mechanics part3 final

+++ app/src/main/java/utenti/diario/utilities/usermanagement/LoginInterface.java
@@ -6,5 +6,5 @@ package utenti.diario.utilities.usermanagement;
 
 public interface LoginInterface {
     void onLoginResult(boolean success, int reason);
-    void onBannedResult(int expiry, String Reason);
+    void onBannedResult(long expiry, String Reason);
 }
